Fix implementation of Storage class and modify spacer length
